+++ controllers/courses.js
@@ -9,19 +9,12 @@ const ErrorResponse = require('../utils/errorResponse');
 //@access Public
 
 exports.getCourses = asyncHandler(async (req,res,next)=>{
-    let query;
     if (req.params.bootcampId){
-        query = Course.find({bootcamp: req.params.bootcampId})
+        const courses = Course.find({bootcamp: req.params.bootcampId})
+        return res.status(200).json({ success:true, count: courses.length, msg: courses});
     } else {
-        query = Course.find().populate({
-            path:'bootcamp',
-            select:'name description',
-            })}
-
-    const courses = await query;
-    res.status(200).json({ success:true, count: courses.length, msg: courses});
-
-})
+        res.status(200).json(res.advancedResults);
+}})
 
 //@desc Get single course
 //@route GET /api/v1/courses/:id
@@ -43,11 +36,16 @@ exports.getCourse = asyncHandler(async (req,res,next)=>{
 //@route POST /api/v1/bootcamps/:bootcampId/courses
 //@access Private
 
-exports.addCourse = asyncHandler(async (req,res) =>{
+exports.addCourse = asyncHandler(async (req,res,next) =>{
     req.body.bootcamp = req.params.bootcampId;  //manually set bootcamp field in Course model to bootcampId from Bootcamp model
+    req.body.user = req.user.id; //manually set user field in Course model to user id from User model
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
     if(!bootcamp){
         next(new ErrorResponse(`No bootcamp with id ${req.params.id}`, 404))
+    }
+    //Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with id ${req.user.id} is not authorized to add a course to ${bootcamp._id}`,401))
     }
     const course = await Course.create(req.body);
     res.status(201).json({success: true, data: course})
@@ -63,6 +61,10 @@ exports.updateCourse = asyncHandler(async(req,res,next) =>{
 
     if (!course){
         return next(new ErrorResponse(`Course with id ${req.params.id} does not exist`, 404))
+    }
+    //Make sure user is course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with id ${req.user.id} is not authorized to update this course`,401))
     }
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -80,6 +82,10 @@ exports.deleteCourse = asyncHandler(async(req,res,next) =>{
 
     if (!course){
         return next(new ErrorResponse(`Course with id ${req.params.id} does not exist`, 404))
+    }
+    //Make sure user is course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User with id ${req.user.id} is not authorized to delete this course`,401))
     }
     await course.deleteOne();
